backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

google-genai==1.75.0

faiss-cpu==1.8.0
PyMuPDF==1.23.6
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.4

sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
email-validator==2.2.0

httpx==0.28.1

pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==7.1.0